Stop returning Any where the dtype is known, and close the rest
Two more mypy tranches. Reading a row out of a grid comes back as ``Any`` from
numpy's stubs, so those returns go through ``np.asarray``; three ``@njit``
scalars are cast where the declared return already said ``float``.

``Any`` that is genuinely ``Any`` now says so at the site: a condition's labels
are whatever pandas holds them as, and an archetype's ``run``/``legs``/``signal``
have a different signature each. ``params_cls`` is ``type[Params]`` rather than
a bare ``type``, which is a real narrowing rather than a suppression.

The trade-log gate is byte-for-byte identical, which is what says the ``float``
casts inside the loops changed nothing.

Part of #56.

+++ nqbt/sim/bracket.py
@@ -318,7 +318,7 @@ def limit_filled(favourable_px: float, limit: float, on_touch: bool, direction: 
 @njit(cache=True)
 def round_to_tick(price: float, tick_size: float) -> float:
     """Snap a price onto the tick grid, as ``RoundToTickSize`` does in NinjaScript."""
-    return np.floor(price / tick_size + 0.5) * tick_size
+    return float(np.floor(price / tick_size + 0.5) * tick_size)
 
 
 @njit(cache=True)
